Resnet module and swish and mish has been transfer to a new file. Module.py
